function/requirements.txt
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
boto3==1.34.12
boto3-type-annotations-with-docs==0.3.1
botocore==1.34.12
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
frozenlist==1.4.1
future==0.18.3
idna==3.6
jmespath==1.0.1
line-bot-sdk==3.7.0
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.4
